+++ include/json/utils.hpp
@@ -6,13 +6,13 @@
 
 namespace GameJson{
 
-void writeProp(std::string camp, int value);
+void writeProp(std::string camp, long long value);
 void writeName(std::string name);
-void writeItem(std::string camp, int value);
+void writeItem(std::string camp, long long value);
 
-int getData(std::string camp);
+long long getData(std::string camp);
 std::string getName();
-int getItem(std::string camp);
+long long getItem(std::string camp);
 
 }
 #endif
