+++ src/game/work.cpp
@@ -1,5 +1,4 @@
-#include "game/balance.hpp"
-#include <cstdlib>
+#include <game/balance.hpp>
 #include <game/work.hpp>
 #include <game/player.hpp>
 #include <json/utils.hpp>
@@ -10,7 +9,7 @@ namespace Economy{
 
 void work(){
     Economy::Balance *bal = new Economy::Balance();
-    int amount = rand() % 500 + 1000 * GameJson::getData("lvl");
+    long long amount = rand() % 500 + 1000 * GameJson::getData("lvl");
     bal->addBalance(amount);
     std::cout << "Has trabajado y ganaste " << amount << "$" << std::endl;
     sleep(1);
